drop old cli and move in new cli
