--- clients/cli2/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "cli2"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-atty = "0.2"
-clap = { version = "4.1.4", features = ["derive"] }
-hotwatch = "0.4.6"
-lb = { path = "../../libs/core", package = "lockbook-core" }
